+++ tomo/src/libs/tomo.type.ts
